requirements.txt
fastapi
uvicorn
pydantic[email]
supabase
loguru
phonenumbers
python-dotenv
httpx
pytest
langgraph
langfuse
langchain-openai
langchain-core
pydantic-settings
celery[redis]
redis
python-multipart
pandas
streamlit
requests

